requirements.txt
# AI/ML Libraries
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langgraph==0.0.20
openai==1.3.0
tiktoken==0.5.2
azure-identity==1.15.0
azure-core==1.29.5

# Vector Database
faiss-cpu==1.7.4
chromadb==0.4.18
sentence-transformers==2.2.2

# Web Framework
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
numpy==1.24.3
yfinance==0.2.18
requests==2.31.0

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Database
sqlite3
sqlalchemy==2.0.23

# Environment & Utils
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Docker
docker==6.1.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
